feat(ui): Support scrolling in preview pane (#176)
Closes #60

This PR adds the ability to scroll the preview pane similar to the diff
scrolling feature. We accomplish this by
setting the history limit when creating the tmux session and then
capturing the full history when the user begins to scroll.
The content is then rendered inside a viewport that is scrollable with
either shift+arrow keys or with the scroll wheel.

Also added some e2e tests on the preview rendering.

<hr />

https://github.com/user-attachments/assets/4fb5762e-29d1-4ae8-8ec4-3db4070890af